+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - autoresume option for Config API ([#907](https://github.com/catalyst-team/catalyst/pull/907))
 - a few issues with TF projector ([#917](https://github.com/catalyst-team/catalyst/pull/917))
 - batch sampler speed issue ([#921](https://github.com/catalyst-team/catalyst/pull/921)) 
+- add apex key-value optimizer support ([#924](https://github.com/catalyst-team/catalyst/pull/924))